Split out Background Manager processes
