+++ db/background_mgr.go
@@ -11,12 +11,10 @@ package db
 import (
 	"net/http"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	sgbucket "github.com/couchbase/sg-bucket"
 	"github.com/couchbase/sync_gateway/base"
-	"github.com/google/uuid"
 	pkgerrors "github.com/pkg/errors"
 )
 
@@ -462,297 +460,4 @@ func (b *BackgroundManager) UpdateHeartbeatDocClusterAware() error {
 
 func (b *BackgroundManager) isClusterAware() bool {
 	return b.clusterAwareOptions != nil
-}
-
-// ======================================================
-// Resync Implementation of Background Manager Process
-// ======================================================
-
-type ResyncManager struct {
-	DocsProcessed int
-	DocsChanged   int
-	lock          sync.Mutex
-}
-
-var _ BackgroundManagerProcessI = &ResyncManager{}
-
-func NewResyncManager() *BackgroundManager {
-	return &BackgroundManager{
-		Process:    &ResyncManager{},
-		terminator: base.NewSafeTerminator(),
-	}
-}
-
-func (r *ResyncManager) Init(options map[string]interface{}, clusterStatus []byte) error {
-	return nil
-}
-
-func (r *ResyncManager) Run(options map[string]interface{}, persistClusterStatusCallback updateStatusCallbackFunc, terminator *base.SafeTerminator) error {
-	database := options["database"].(*Database)
-	regenerateSequences := options["regenerateSequences"].(bool)
-
-	defer atomic.CompareAndSwapUint32(&database.State, DBResyncing, DBOffline)
-	callback := func(docsProcessed, docsChanged *int) {
-		r.lock.Lock()
-		defer r.lock.Unlock()
-		r.DocsProcessed = *docsProcessed
-		r.DocsChanged = *docsChanged
-	}
-
-	_, err := database.UpdateAllDocChannels(regenerateSequences, callback, terminator)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (r *ResyncManager) ResetStatus() {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-
-	r.DocsProcessed = 0
-	r.DocsChanged = 0
-}
-
-type ResyncManagerResponse struct {
-	BackgroundManagerStatus
-	DocsChanged   int `json:"docs_changed"`
-	DocsProcessed int `json:"docs_processed"`
-}
-
-func (r *ResyncManager) GetProcessStatus(backgroundManagerStatus BackgroundManagerStatus) ([]byte, error) {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-
-	retStatus := ResyncManagerResponse{
-		BackgroundManagerStatus: backgroundManagerStatus,
-		DocsChanged:             r.DocsChanged,
-		DocsProcessed:           r.DocsProcessed,
-	}
-
-	return base.JSONMarshal(retStatus)
-}
-
-// =====================================================================
-// Tombstone Compaction Implementation of Background Manager Process
-// =====================================================================
-
-type TombstoneCompactionManager struct {
-	PurgedDocCount int64
-}
-
-var _ BackgroundManagerProcessI = &TombstoneCompactionManager{}
-
-func NewTombstoneCompactionManager() *BackgroundManager {
-	return &BackgroundManager{
-		Process:    &TombstoneCompactionManager{},
-		terminator: base.NewSafeTerminator(),
-	}
-}
-
-func (t *TombstoneCompactionManager) Init(options map[string]interface{}, clusterStatus []byte) error {
-	database := options["database"].(*Database)
-	database.DbStats.Database().CompactionAttachmentStartTime.Set(time.Now().UTC().Unix())
-
-	return nil
-}
-
-func (t *TombstoneCompactionManager) Run(options map[string]interface{}, persistClusterStatusCallback updateStatusCallbackFunc, terminator *base.SafeTerminator) error {
-	database := options["database"].(*Database)
-
-	defer atomic.CompareAndSwapUint32(&database.CompactState, DBCompactRunning, DBCompactNotRunning)
-	callback := func(docsPurged *int) {
-		atomic.StoreInt64(&t.PurgedDocCount, int64(*docsPurged))
-	}
-
-	_, err := database.Compact(true, callback, terminator)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-type TombstoneManagerResponse struct {
-	BackgroundManagerStatus
-	DocsPurged int64 `json:"docs_purged"`
-}
-
-func (t *TombstoneCompactionManager) GetProcessStatus(backgroundManagerStatus BackgroundManagerStatus) ([]byte, error) {
-	retStatus := TombstoneManagerResponse{
-		BackgroundManagerStatus: backgroundManagerStatus,
-		DocsPurged:              atomic.LoadInt64(&t.PurgedDocCount),
-	}
-
-	return base.JSONMarshal(retStatus)
-}
-
-func (t *TombstoneCompactionManager) ResetStatus() {
-	atomic.StoreInt64(&t.PurgedDocCount, 0)
-}
-
-// =====================================================================
-// Attachment Compaction Implementation of Background Manager Process
-// =====================================================================
-
-type AttachmentCompactionManager struct {
-	MarkedAttachments base.AtomicInt
-	PurgedAttachments base.AtomicInt
-	CompactID         string
-	Phase             string
-	dryRun            bool
-	lock              sync.Mutex
-}
-
-var _ BackgroundManagerProcessI = &AttachmentCompactionManager{}
-
-func NewAttachmentCompactionManager(bucket base.Bucket) *BackgroundManager {
-	return &BackgroundManager{
-		Process: &AttachmentCompactionManager{},
-		clusterAwareOptions: &ClusterAwareBackgroundManagerOptions{
-			bucket:        bucket,
-			processSuffix: "compact",
-		},
-		terminator: base.NewSafeTerminator(),
-	}
-}
-
-func (a *AttachmentCompactionManager) Init(options map[string]interface{}, clusterStatus []byte) error {
-
-	database := options["database"].(*Database)
-	database.DbStats.Database().CompactionAttachmentStartTime.Set(time.Now().UTC().Unix())
-
-	newRunInit := func() error {
-		uniqueUUID, err := uuid.NewRandom()
-		if err != nil {
-			return err
-		}
-
-		dryRun, _ := options["dryRun"].(bool)
-		if dryRun {
-			base.Infof(base.KeyAll, "Attachment Compaction: Running as dry run. No attachments will be purged")
-		}
-
-		a.dryRun = dryRun
-		a.CompactID = uniqueUUID.String()
-		base.Infof(base.KeyAll, "Attachment Compaction: Starting new compaction run with compact ID: %q", a.CompactID)
-		return nil
-	}
-
-	if clusterStatus != nil {
-		var attachmentResponseStatus AttachmentManagerResponse
-		err := base.JSONUnmarshal(clusterStatus, &attachmentResponseStatus)
-
-		reset, ok := options["reset"].(bool)
-		if reset && ok {
-			base.Infof(base.KeyAll, "Attachment Compaction: Resetting compaction process. Will not  resume any "+
-				"partially completed process")
-		}
-
-		// If the previous run completed, or there was an error during unmarshalling the status we will start the
-		// process from scratch with a new compaction ID. Otherwise, we should resume with the compact ID, phase and
-		// stats specified in the doc.
-		if attachmentResponseStatus.State == BackgroundProcessStateCompleted || err != nil || (reset && ok) {
-			return newRunInit()
-		} else {
-			a.CompactID = attachmentResponseStatus.CompactID
-			a.Phase = attachmentResponseStatus.Phase
-			a.dryRun = attachmentResponseStatus.DryRun
-			a.MarkedAttachments.Set(attachmentResponseStatus.MarkedAttachments)
-			a.PurgedAttachments.Set(attachmentResponseStatus.PurgedAttachments)
-
-			base.Infof(base.KeyAll, "Attachment Compaction: Attempting to resume compaction with compact ID: %q phase %q", a.CompactID, a.Phase)
-		}
-
-		return nil
-
-	}
-
-	return newRunInit()
-}
-
-func (a *AttachmentCompactionManager) Run(options map[string]interface{}, persistClusterStatusCallback updateStatusCallbackFunc, terminator *base.SafeTerminator) error {
-	database := options["database"].(*Database)
-
-	persistClusterStatus := func() {
-		err := persistClusterStatusCallback()
-		if err != nil {
-			base.Warnf("Failed to persist cluster status on-demand following completion of phase: %v", err)
-		}
-	}
-
-	defer persistClusterStatus()
-
-	// Need to check the current phase in the event we are resuming - No need to run mark again if we got as far as
-	// cleanup last time...
-	switch a.Phase {
-	case "mark", "":
-		a.SetPhase("mark")
-		persistClusterStatus()
-		_, err := Mark(database, a.CompactID, terminator, &a.MarkedAttachments)
-		if err != nil || terminator.IsClosed() {
-			return err
-		}
-		fallthrough
-	case "sweep":
-		a.SetPhase("sweep")
-		persistClusterStatus()
-		_, err := Sweep(database, a.CompactID, a.dryRun, terminator, &a.PurgedAttachments)
-		if err != nil || terminator.IsClosed() {
-			return err
-		}
-		fallthrough
-	case "cleanup":
-		a.SetPhase("cleanup")
-		persistClusterStatus()
-		err := Cleanup(database, a.CompactID, terminator)
-		if err != nil || terminator.IsClosed() {
-			return err
-		}
-	}
-
-	a.SetPhase("")
-	return nil
-}
-
-func (a *AttachmentCompactionManager) SetPhase(phase string) {
-	a.lock.Lock()
-	defer a.lock.Unlock()
-
-	a.Phase = phase
-}
-
-type AttachmentManagerResponse struct {
-	BackgroundManagerStatus
-	MarkedAttachments int64  `json:"marked_attachments"`
-	PurgedAttachments int64  `json:"purged_attachments"`
-	CompactID         string `json:"compact_id"`
-	Phase             string `json:"phase,omitempty"`
-	DryRun            bool   `json:"dry_run,omitempty"`
-}
-
-func (a *AttachmentCompactionManager) GetProcessStatus(status BackgroundManagerStatus) ([]byte, error) {
-	a.lock.Lock()
-	defer a.lock.Unlock()
-
-	retStatus := AttachmentManagerResponse{
-		BackgroundManagerStatus: status,
-		MarkedAttachments:       a.MarkedAttachments.Value(),
-		PurgedAttachments:       a.PurgedAttachments.Value(),
-		CompactID:               a.CompactID,
-		Phase:                   a.Phase,
-		DryRun:                  a.dryRun,
-	}
-
-	return base.JSONMarshal(retStatus)
-}
-
-func (a *AttachmentCompactionManager) ResetStatus() {
-	a.lock.Lock()
-	defer a.lock.Unlock()
-
-	a.MarkedAttachments.Set(0)
-	a.PurgedAttachments.Set(0)
-	a.dryRun = false
 }
